Rewrite the simulation graph component.  It is now implemented as a fragment with a menu to configure the serieses and display a list of the flight events.
SimulationChart now implements a flyweight pattern containing the simulationId so it can be serialized allows the os to pause the application and restore most of the state of the graph.  Currently the selected serieses are not persisted so the graph gets reset to the default.

The simulation view is operated in two different modes.  On tablets in landscape the graph appears in the right sidepane and the list on the left.  When a new simulation is selected, the old graph is pushed to the back stack.  When in portrait mode or on smaller sceens, the simluation graph is in a separate activity.