De-localize Material and FlightDataType objects when persisting in ork files.  Added a String key to each of these objects which holds on to a unique identifier which is stored in the ork file.  User defined Materials are given a key which is defined as "UserDefined." + name.
The construction pattern for Material has changed so explicitly passing in the userDefined boolean is no longer required.  Instead there are two different factory methods to be used.  Direct construction of Material derived classes is no longer allowed.

The persistence mechanism for Materials in perferences has changed.  It will now persist the key and name but still supports reading old format.

The example files have been regenerated, but will need to be regenerated again prior to an official build in order to get the version string correct.