Fix problematic bug in motor selection which manifested as:
java.lang.IllegalArgumentException: setSelectedIndex: 0 out of bounds
	at javax.swing.JComboBox.setSelectedIndex(Unknown Source)
	at
net.sf.openrocket.gui.dialogs.motor.thrustcurve.ThrustCurveMotorSelectionPanel.updateData(ThrustCurveMotorSelectionPanel.java:437)

The underlying problem was the MotorChooserDialog had been using the
GUIUtil.setDisposableDialogOptions call which caused all the underlying
models to be cleared when the dialog was closed with the "x" button.
This caused the reference to the models in the panel to be disassociated
with those in the swing widgets.

Since the dialog is reused, this behavior was bad.  Changed the dialog
so the 'x' button simply hides the dialog and use the
GUIUtil.installEscapeCloseOperation to wire in the ESC key.