Modified setTabHeight, setTabLength, and setTabOffset to take a second parameter, determining whether to validate fin tab, and made them private.
Created two public wrappers for each, one calling for validation
(which has the same signature as the original three methods), and the
other not.

Called NoValidate versions of the three in file loading functions so
validation doesn't interfere with tab length when tab is read before
fin.

Tried to make the coding of the three more consistent with each
other.