Add snapcraft information for building snaps
Snaps are a universal packaging format for Linux platforms
and is available for most major Linux distributions.

This commit adds support for building OpenRocket as a snap. The
snap convention is to place snap relevant information in the top
level snap directory. The snapcraft.yaml file provides information
regarding what should be in the snap and how to build it.

Signed-off-by: Billy Olsen <billy.olsen@gmail.com>