Update rail button and tube fin icons for transparency
Update the transparency of the exported icons so there isn't a white
background clashing with the button color. Additionally, the tube fin
and rail button small icons were updated to be more in line with other
icons.

The tube fin is bigger and has the body tube showing like other
fin small icons.

The rail button works better in a side profile rather
than an offset profile.

Signed-off-by: Billy Olsen <billy.olsen@gmail.com>