Do not show graphics editor options from snap
Gracefully handle the graphics editor options when running
from within a snap. When running in snap confinement, the graphics
editor option in the preferences panel is removed and edits to an
existing decal will use xdg-open, allowing the user to select the
appropriate application for editing using the system options.

Signed-off-by: Billy Olsen <billy.olsen@gmail.com>