Don't set a warning if recovery device is deployed (for instance, due to apogee detect) while motor is coasting.
Background:  a motor that has burned out can be in one of two states:  DELAYING (the delay charge has not yet fired) or SPENT (either the delay charge, if any, has fired or there was no delay charge).

The existing SimulationStatus:getActiveMotors() method returned a list of motors that were active for the stage, and which were not SPENT.  The test for the warning redudantly tested for a SPENT motor, still didn't test for a DELAYING motr, and consequently set the warning.

This PR
(1) adds a boolean MotorClusterStatus:isDelaying() method, analogous to the existing state test methods

(2) modifies SimulationStatus.java:getActiveMotors() to return all the motors, not just the ones that aren't SPENT.  This is to improve consistency with FlightConfiguration:getActiveMotors(), and to make for a more consistent usage between the two calls to the method.

(3) adds !isDelaying() to the test for the warning.