Adjust grid in freeform fin editor
Adjust the colors in the freeform fin editor to use a blue hue as
was used in previous releases. Major grid lines are slightly darker
than minor grid lines. Also Adjust selected point color to be red.

Additionally, adjust the minor and major tick size to scale with the
current selection.

Fixes #836
Fixes #837

Signed-off-by: Billy Olsen <billy.olsen@gmail.com>