Add extra spacing for stability information text
The GlyphVector.getVisualBounds() does not include leading or trailing
spaces. Account for this when drawing stability information by
increasing the unitwidth by the length of one space.

Additionally, the code is inefficient in that it creates a new Font
object each and every time a new text string is created for graphical
display. This is unnecessary and adds overhead to the VM memory
requirements.

Fixes #880

Signed-off-by: Billy Olsen <billy.olsen@gmail.com>