Fix transparency display issue
Use the provided alpha from the color when it is available when
rendering the model in the 3d finished view. The 3d unfinished view will
continue to override the alpha to show interior components.

Additionally, enable the GL2.GL_COLOR_MATERIAL for the rendering of
the component to display the right alpha settings.

Fixes #878

Signed-off-by: Billy Olsen <billy.olsen@gmail.com>