Invalidate MotorInformationPanel when updating (fixes #885)
Calls invalidate() as the last action inside MotorInformationPanel.updateData(), forcing a redraw of the panel.  This makes sure the panel is large enough to display all of its contents

Also, clean up display of the compatible cases a little bit by adding a space between cases, and wrapping in <html></html> tags to allow line breaks.