Make spinner options in photo keyboard editable
Spinner options in photo frame does not have an editor, so doesn't
allow for keyboard edits. Add an EditableSpinner class which
automatically sets the editor and use that instead of JSpinner.

Note: this EditableSpinner can be used throughout OR, but not including
everywhere as it is a large change at this point and I'd rather wait for
the refactor until after release, so change is small.

Fixes #873

Signed-off-by: Billy Olsen <billy.olsen@gmail.com>