Update license dialog Quit quoting entire GPL in dialog; insert hrefs pointing to it. Add other licenses:
Apache license, for Dave's components library

DejaVu font license.  I wasn't able to find an "authoritative" source
for this license to point people at.

BSD 2-Clause license for commonmark-java library.  This license is so short I
left it in the dialog.  To my surprise, I also didn't find an "authoritative"
source for this license to point people at