Update snapcraft builds for dbcook's db
The update to use dbcook's component database adds a git submodule. The
submodule configuration needs to use an https transport rather than an
ssh transport for users not logged in through git (snapcraft builds
aren't; they're in a clean room).

Additionally, the snapcraft.yaml file needs to be told the source type
is git in order to automatically udpate an pull in submodules.

Signed-off-by: Billy Olsen <billy.olsen@gmail.com>