Make motor database search more robust, especially when loading files:
Instead of requiring motor "designation" to match either the actual designation or the common name, require it to be a substring of the actual designation or that the common name be a substring of the actual designation.  Many motors in stored rocket files have "designations" that resemble, but do not match, actual designations and common names.

Don't match on diameter or length when loading files.  Motors with errors in the old database (eg M1850) would fail to be recognized.