Revised presets to allow .orc files to contain parachute spill hole diameter and surface area data fields
A large number of the new parachute products have spill holes (c.f. Issue #817), and simply adding those fields to the .orc file, without changes to how the presets are handled, was causing OpenRocket to throw errors. These changes allow these fields to not exists or  to exist with or without data, without throwing an error. This also means that the parachute .orc files will not need to be updated when spill holes are implemented.