Mark the snap grade as stable
The grade of the snap is marked as devel, which is great while the
app is being developed but means that the snap is not eligible to be
published to candidate or stable release channels. This change marks
the snap grade as stable to allow it to be released.

Additionally, the automatic snap publishing that happens for each
commit pushed to the unstable branch has been failing to release
the snap revisions due to a symlink introduced in the openjdk package
for which leaves a dangling symlink certs. The blacklisted.certs
should be now contained in blocked.certs so we add that to the list
of files to remove when being primed (final preparation for snap
packaging).

Signed-off-by: Billy Olsen <billy.olsen@gmail.com>