Got rid of black background and flame
OK, I got rid of the flame because... well, not for a good reason actually, but afterwards I realized that including the flame is tricky if you want the rocket to fill the available space.  I can always put it back, but we might as well try this one on for size since it's already done.