Use current atmospheric conditions, not previous conditions, in calculating thrust.
This lets us get rid of previousAtmosphericConditions completely.

Note that while we pass acceleration and atmospheric conditions into calculateAverageThrust(), we don't actually do anything with them there. They appear to be passed in for some sort of eventual enhancements when we do things like altitude compensation for motors, or acceleration for hybrids or water rockets.

It also turned out that while we called getPreviousAtmosphericConditions() when call ing calculateAverageThrust(), setPreviousAtmosphericConditions() was never called.