Remove longitudinalAcceleration field from DataStore.
This was initialized to NaN, but never set to anything else. It (along with the atmospheric conditions) was passed to calculateThrust for possible more accurate thrust calculation in the future, but not used there at present

The entire DataStore is now passed to calculateAcceleration() to use whatever fields are useful to it -- none at present.