Load and store SIM_WARN flight events
This requires substantial changes to Warnings in .ork files. Instead of Warnings consisting of a text string with id and priority as attributes, it is now a mixed content element with id, priority, description, and possibly sources as subelements and the original text string still present to provide backward compatibility.

If a .ork file with a warning is saved with this PR and then loaded also with this PR, the warning is reconstructed from the subelements and is available if there is a SIM_WARN flight event. If at some point we want to highlight the components referenced in a Warning, we now have references to them.

If a .ork file with a warning is saved with 23.09 and then loaded with this PR, none of the subelements are present, so the Warning is reconstructed from the text.  It won't have actual references to the components mentioned in the Warning, and there won't be a SIM_WARN event needing reference the Warning.

If a .ork file with a warning is saved with this PR and then loaded in 23.09, there will be .ork loading warnings for all the elements defined in this PR, and then the Warning will be constructed from the text content just as for a .ork saved in 23.09.