Update flight event unit tests
Center booster no longer has the "open front" warning (the whole point of this exercise!).
Made the test motors a closer approximation to the actual Estes A10 and C6 motors so the rocket will be fast enough to trigger a "recovery at high speed" warning.
Added parachutes to the side boosters with recovery at motor burnout. This gives a high speed deployment warning so warnings are still tested.
Took the fins off the sustainer so it won't be stable; this tests simulation abort on "tumble under thrust".
Took out some unnecessary logging.