Instead of storing CP (with CNa as a weight) store the moment defined by CP * CNa.
When mergine AerodynamicForces, this lets us just add the moments instead of doing a weight average of the CPs; this in turn avoids a divide-by-zero error caused by nose cones having a weight of 2 and tail cones having a weight of -2 when accumulating CP for a rocket.

setCP() still expects a Coordinate consisting of the CP with a weight defined by CNa, and getCP() still returns a similar Coordinate.

When setCP() is given a CP with a weight of 0, the x, y, z values are replaced by 0. When getCP() is called and the weight is 0, the CP values returned are 0 (it would arguably be more correct to return NaN, that could be a confusing result for really tiny fins).